feat: nav group
Add a new feature to the nav group, which is a collection of nav items. In wide viewports, the nav group is displayed as a dropdown menu when the user hover over the nav group title. In narrow viewports, the nav group is displayed as a submenu in drawer menu when the user click on the nav group title.